Update template for 0.6.0
The API is changing so our default template needs some adjustments.
There are also a small number of inconsistencies to the template that we
can address at the same time.

First and foremost, this simply adjusts the intent of the template to
use the new API.

Second, to enable URL purging, apps must accept and forward "PURGE"
requests, so this now does so.

But last, the routing logic needs a little bit of cleanup.  While one
section of the code suggested that HEAD and POST requests were accepted,
they would later all be issued 404 responses.  The vast majority of
Fastly services have no real need to support POST requests, so we
shouldn't imply they should in our blessed example templates.

Differentiating GETs and HEADs can potentially get tricky, but in this
simple example, we may simply treat them as though they are the same.
If a Response with a body is made to a HEAD request, the body will be
dropped, but its Content-Length will be sent.